requirements.txt
discord.py==2.5.2
aiohttp==3.11.18
python-dotenv==1.1.0
pytest==8.3.5
